reduce example added
