+++ cpp/CMakeLists.txt
@@ -37,3 +37,4 @@ endfunction()
 create_target(hello_world "src/hello_world.cpp")
 create_target(send_recv "src/send_recv.cpp")
 create_target(round_robin "src/round_robin.cpp")
+create_target(reduce "src/reduce.cpp")
